# docs/rest_v1_openapi.yaml
openapi: 3.0.2
servers:
  - url: http://localhost:3370
    description: Local Linstor controller
info:
  description: |
    Linstor REST API V1

    The V1 rest api of Linstor should stay compatible and only additions are made to the API,
    If there are breaking changes or redesigned a new major REST API version will be issued.

    Server runs per default on port `3370` on `::` ipv6 and ipv4.

    To change the bind address or port you can use the following linstor client commands:
    ```
    linstor controller set-property REST/bindAddress 127.0.0.1
    linstor controller set-property REST/port 8080
    ```

    After setting this properties restart the controller and the new values should be used.

    Changelog:

    * unreleased
       - Added makeResourceAvailable API
       - Added diskless storage pool filter to AutoSelectFilter
       - Added external_locking to StoragePool
    * 1.6.0
       - Added PropsInfoAll, PropsInfo and PropsInfoDetails as well as these entrypoints:
          - /v1/controller/properties/info
          - /v1/controller/properties/info/all
          - /v1/nodes/properties/info
          - /v1/storage-pool-definitions/properties/info
          - /v1/nodes/{node}/storage-pools/properties/info
          - /v1/resource-definitions/properties/info
          - /v1/resource-definitions/{resource}/resources/properties/info
          - /v1/resource-definitions/{resource}/volume-definitions/properties/info
...
       - Added /v1/nodes/{node}/restore
...
paths:
...
  /v1/nodes/{node}/storage-pools:
    parameters:
      - $ref: '#/components/parameters/Node'
    get:
      tags:
        - developers
      summary: lists all storage pools for a node
      operationId: nodeStoragePoolList
      description: Lists all storage pools of a node
      parameters:
        - $ref: '#/components/parameters/Nodes'
        - name: storage_pools
          in: query
          description: Filter only for the specified storage pools, if not specified
            no filtering.
          schema:
            type: array
            items:
              type: string
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Array list of all storage pools
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StoragePool'
    post:
      tags:
        - developers
      summary: creates a new storage pool on this node
      operationId: nodeStoragePoolCreate
      description: |
        Creates a new storage pool on this node.

        `provider_kind` has to be specified and additional to that
        the linked `StorDriver` property that has the value to the backing pool:

          * `LVM`: `StorDriver/LvmVg`
          * `LVM_THIN`: `StorDriver/LvmVg` and `StorDriver/ThinPool`
          * `ZFS`: `StorDriver/ZPool`
          * `ZFS_THIN`: `StorDriver/ZPoolThin`
          * `DISKLESS`: Does not need a property as it has no backing pool
      responses:
        '201':
          description: storage pool created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiCallRcList'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StoragePool'
            example:
              storage_pool_name: DfltStorPool
              provider_kind: LVM_THIN
              props:
                StorDriver/LvmVg: mylvmpool
                StorDriver/ThinPool: thin
  /v1/nodes/{node}/storage-pools/{storagepool}:
    parameters:
      - $ref: '#/components/parameters/Node'
      - $ref: '#/components/parameters/StoragePool'
    put:
      tags:
        - developers
      summary: modify a storage pool
      operationId: nodeStoragePoolModify
      description: |
        Sets or modifies properties

        Possible properties are:
        - `StorDriver/StorPoolName` - regex[`[a-zA-Z0-9_/-]+`]
        - `StorDriver/LvmVg` - regex[`[a-zA-Z0-9_-]+`]
        - `StorDriver/ThinPool` - regex[`[a-zA-Z0-9_-]+`]
        - `StorDriver/ZPool` - regex[`[a-zA-Z0-9_/-]+`]
        - `StorDriver/ZPoolThin` - regex[`[a-zA-Z0-9_/-]+`]
        - `StorDriver/FileDir` - regex[`.*`]
        - `PrefNic` - regex[`^[a-zA-Z0-9_][a-zA-Z0-9_-]{2,47}$`]

            Preferred network interface to use

        - `NVMe/PrefNic` - regex[`^[a-zA-Z0-9_][a-zA-Z0-9_-]{2,47}$`]

            Preferred network interface to use

        - `StorDriver/LvcreateType` - enum
            * linear
            * striped
            * mirror
            * raid0
            * raid1
            * raid4
            * raid5
            * raid6
            * raid10
            * lzma
            * lz4

        - `StorDriver/LvcreateOptions` - regex[`.*`]

            Additional parameters added to every 'lvcreate ... ' command

        - `StorDriver/ZfscreateOptions` - regex[`.*`]

            Additional parameters added to every 'zfs create ... ' command

        - `StorDriver/WaitTimeoutAfterCreate` - regex[`[0-9]+`]
        - `sys/fs/blkio_throttle_read` - long

            Sets the /sys/fs/cgroup/blkio/blkio.throttle.read_bps_device

        - `sys/fs/blkio_throttle_write` - long

            Sets the /sys/fs/cgroup/blkio/blkio.throttle.write_bps_device

        - `sys/fs/blkio_throttle_read_iops` - long

            Sets the /sys/fs/cgroup/blkio/blkio.throttle.read_iops_device

        - `sys/fs/blkio_throttle_write_iops` - long

            Sets the /sys/fs/cgroup/blkio/blkio.throttle.write_iops_device

        - `StorDriver/Openflex/ApiHost` - regex[`.+`]

            Openflex API host name

        - `StorDriver/Openflex/ApiPort` - regex[`[0-9]+`]

            Openflex API port

        - `StorDriver/Openflex/UserName` - regex[`.+`]

            Openflex user name

        - `StorDriver/Openflex/UserPassword` - regex[`.+`]

            Openflex password

        - `StorDriver/Openflex/StorDev` - regex[`.+`]

            Openflex storage device

        - `StorDriver/Openflex/StorDevHost` - regex[`.+`]

            Openflex storage device host

        - `StorDriver/Openflex/StorPool` - regex[`[0-9]+`]

            Openflex storage pool name

        - `Autoplacer/MaxThroughput` - long

            The maximum throughput the given storage pool is capable of.

      responses:
        '200':
          description: storage pool modified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiCallRcList'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StoragePoolDefinitionModify'
            example:
              override_props:
                Aux/testprop: '1'
              delete_props:
                - Aux/deleteprop
    delete:
      tags:
        - developers
      summary: delete a storage pool
      operationId: nodeStoragePoolDelete
      description: Delete a storage pool
      responses:
        '200':
          description: storage pool deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiCallRcList'
  /v1/nodes/{node}/storage-pools/properties/info:
    parameters:
      - $ref: '#/components/parameters/Node'
    get:
      summary: list all storage-pool properties
      description: list all property-names and their descriptions
      responses:
        '200':
          description: list of property-names
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  $ref: '#/components/schemas/PropsInfo'
...
  /v1/view/storage-pools:
    get:
      tags:
        - developers
      operationId: viewStoragePools
      summary: query all storage pools
      description: This REST-resource should be used if you want to get an overview
        of all storage pools.
      parameters:
        - $ref: '#/components/parameters/Nodes'
        - name: storage_pools
          in: query
          description: Filter only for the specified storage pools, if not specified
            no filtering.
          schema:
            type: array
            items:
              type: string
        - $ref: '#/components/parameters/Props'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: storage pool list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StoragePool'
        '500':
          $ref: '#/components/responses/OperationFailed'
...
components:
  parameters:
...
    Node:
      name: node
      in: path
      required: true
      description: node to use
      schema:
        type: string
    Nodes:
      name: nodes
      in: query
      description: Filter only for the specified nodes, if not specified, no filtering.
      schema:
        type: array
        items:
          type: string
...
    StoragePool:
      name: storagepool
      in: path
      required: true
      description: Storage pool to use
      schema:
        type: string
    Offset:
      name: offset
      in: query
      description: number of records to skip for pagination
      schema:
        type: integer
        format: int32
        minimum: 0
    Limit:
      name: limit
      in: query
      description: maximum number of records to return
      schema:
        type: integer
        format: int32
        minimum: 0
        maximum: 50
    Props:
      name: props
      in: query
      description: filter by given properties, full property path
      schema:
        type: array
        items:
          type: string
...
  responses:
...
    OperationFailed:
      description: operation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiCallRc'
  schemas:
    ApiCallRc:
      description: Common api reply structure
      type: object
      required:
        - ret_code
        - message
      properties:
        ret_code:
          type: integer
          format: int64
          description: A masked error number
        message:
          type: string
          example: Operation result.
        cause:
          type: string
          description: Cause of the error
        details:
          type: string
          description: Details to the error message
        correction:
          type: string
          description: Possible correction options
        error_report_ids:
          type: array
          description: List of error report ids related to this api call return code.
          items:
            type: string
        obj_refs:
          type: object
          description: Map of objection that have been involved by the operation.
          additionalProperties:
            type: string
    ApiCallRcList:
      type: array
      items:
        $ref: '#/components/schemas/ApiCallRc'
...
    Properties:
      type: object
      description: A string to string property map.
      additionalProperties:
        type: string
    ProviderKind:
      type: string
      enum:
        - DISKLESS
        - LVM
        - LVM_THIN
        - ZFS
        - ZFS_THIN
        - OPENFLEX_TARGET
        - FILE
        - FILE_THIN
        - SPDK
    StoragePool:
      description: |
        Contains information about a storage pool.

        If state is `Error` check the storage pool object path for detailed error description:

          /v1/nodes/{nodename}/storage-pools/{poolname}
      type: object
      required:
        - storage_pool_name
        - provider_kind
      properties:
        storage_pool_name:
          type: string
          example: DfltStorPool
        node_name:
          type: string
        provider_kind:
          $ref: '#/components/schemas/ProviderKind'
        props:
          $ref: '#/components/schemas/Properties'
        static_traits:
          type: object
          readOnly: true
          description: read only map of static storage pool traits
          additionalProperties:
            type: string
        free_capacity:
          type: integer
          format: int64
          example: 882894848
          description: Kibi - read only
        total_capacity:
          type: integer
          format: int64
          example: 1056964608
          description: Kibi - read only
        free_space_mgr_name:
          type: string
          example: drbd1:myfastpool
          description: read only
        uuid:
          type: string
          description: unique object id
          example: e8ef8d6b-17bc-42f0-9367-4aae40c78ecb
        reports:
          type: array
          description: Currently known report messages for this storage pool
          readOnly: true
          items:
            $ref: '#/components/schemas/ApiCallRc'
        supports_snapshots:
          type: boolean
          description: true if the storage pool supports snapshots. false otherwise
        external_locking:
          type: boolean
          description: true if a shared storage pool uses linstor-external locking, like cLVM
...
    PropsInfo:
      type: object
      properties:
        info:
          type: string
        prop_type:
          type: string
        value:
          type: string
        dflt:
          type: string
        unit:
          type: string
...
    StoragePoolDefinitionModify:
      type: object
      properties:
        override_props:
          $ref: '#/components/schemas/Properties'
        delete_props:
          type: array
          items:
            type: string
        delete_namespaces:
          type: array
          items:
            type: string
...